+++ rector.php
@@ -2,19 +2,23 @@
 
 declare(strict_types=1);
 
+use Contao\Rector\Set\ContaoLevelSetList;
+use Rector\CodeQuality\Rector\If_\CombineIfRector;
 use Rector\Config\RectorConfig;
-use Rector\Set\ValueObject\LevelSetList;
 use Rector\Set\ValueObject\SetList;
 
 return RectorConfig::configure()
     ->withPaths([
         __DIR__ . '/src',
     ])
+    ->withSkip([
+        CombineIfRector::class
+    ])
     ->withSets([
         SetList::CODING_STYLE,
         SetList::CODE_QUALITY,
-        LevelSetList::UP_TO_PHP_80,
-        LevelSetList::UP_TO_PHP_81,
-        LevelSetList::UP_TO_PHP_82,
-        SetList::DEAD_CODE
+        SetList::DEAD_CODE,
+        SetList::TYPE_DECLARATION,
+        ContaoLevelSetList::UP_TO_CONTAO_53,
+
     ]);